requirements.txt
langchain==0.0.149
python-dotenv==1.0.0
tiktoken==0.3.3
chromadb==0.3.21
